Enable 'runCheckingPasses' in all soy unit tests that use SoyFileSetParserBuilder.
This is in preparation for moving all those passes into ParsePasses.  The vast majority of the changes in this CL are adding param declarations to templates.  The rest are:
* Working around 'cannot print bool' errors
* Fixing v1 expr syntax in a few places

This also makes the insanity of some of the tests more obvious.  Of note
ResolveNamesVisitor.testNameLookupFailure : which is testing a feature that is _disabled_ in ResolveNamesVisitor but is actually implemented by CheckTemplateParamsVisitor
CheckTemplateParamsVisitorTest.testCallWithMissingParam which checks that missing params aren't errors, when in fact they are.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=103133453