Soy jbcsrc: Stop requiring a default message bundle for SoySauce instead store default messages as constant in the classes.
Currently SoySauce requires that users supply a default bundle in addition to a locale specific bundle in order for {msg} nodes to work in development mode (when there is no translation available) and to provide english language fallbacks.  This is not a great solution because it requires users to pass an additional msg bundle and they may not have it, currently we work around this by Extracting messages in SoyFileSet.  However this solution won't work in a precompilation usecase so instead we are moving default messages to static fields in the generated classes.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=105901986