Add fully qualified deltemplate names to the parseinfos
This will make it possible to construct a deltemplate selector for soysauce
without needing to run the compiler. The idea is that the user will hand us a
set of SoyFileInfos and we can use that to find all the public basic templates
and all the deltemplates.  The combination of these two things will allow us to
* construct a deltemplate selector (see []
* calculate transitive used ij params (see []

With these two things, we can construct a SoySauce object from classes found on
the normal classpath.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=106339837