Soy jbcsrc: Add information about used injected parameters and callees to @TemplateMetadata
This is a prerequisite for adding precompilation support.  The SoySauce object has a method getTransitiveIjParamsForTemplate which is currently implemented via a specific compiler pass.  In order to support this method in a precompilation usecase we will need to store this information somewhere else.  The current plan is to add metadata to each generated template and then lazily calculate this data by crawling through the callgraph.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=106638022