Support calls to text templates in HTML for Incremental DOM code generation
Wrap calls to text templates in an HTML context with a call to itext so that a Text node is generated with the return value of text yielding templates. HTML templates modify the tree in place so no additional work needs to be done.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=106737813