Support attributes with no value if they appear in a Soy node.
If encountering the end of a RawTextNode while building an attribute name, treat it as if there had been a space at the end. This does not match the semantics of any other backend, but that would be very annoying to support. An example of a situation where the Incremental DOM code gen would fail is for the case where you have something like:

{if $condition}data-{/if}{if $condition}foo{/if}=bar

This change adds support for the much more common situation would be one where you have something like:

{if $isDisabled}disabled{/if}
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=107149058