Refactor Soys guice modules to simplify the interface and make it possible for code that doesn't need the compiler to not install compiler passes and introduce a new Module for configuring the precompiled SoySauce
In particular,
c.g.template.soy.SoyModule is used to configure the compiler.  In the long run we should seek to remove bindings from this module.
c.g.template.soy.shared.internal.SharedModule  is shared between the compiler and the soysauce runtime.  It installs plugins and the apicallscope.
c.g.template.soy.jbcsrc.api.PrecompiledSoyModule uses SharedModule and _not_ SoyModule and can be used to construct a SoySauce object from precompiled classes

PrecompiledModule only needs the user to supply the set of deltemplates.
Currently this is done via an optional guice binding.  I have also considered
only exposing a Factory object that the user would need to explicitly invoke.
They each have their pros and cons and for now at this this api is not in the
opensource release (so changing our minds about it should be easy in the future).
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=108051696