Fix a bug where blessStringAsTrustedResourceUrlForLegacy would disable all escaping.
The compiler had this incorrect assumption that all directives that have an
EscapingMode will also disable autoescape.  What happened is the bless
directive didn't trigger the compiler error that you can't use
autoescape-cancelling directives. But it also cancelled autoescape.

This CL is just a quick fix, and it's hard to tell what's going on here.  I'm
going to follow this up with a CL to make InferenceEngine more robust -- make
it so that inferred and explicit directives are separate.  If this isn't clear.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=109129580