Favor syntax error over IllegalStateException for duplicate name.
It seems to me that having a duplicate definition of a template is more a
syntax error than an illegal state.  It would be nice to communicate this
back to the template creator in the form of a SoySyntaxException which
will at least point them to the duplicate template definition.

This likely only applies to template files having:

autoescape="deprecated-noncontextual"