attempt to fix the build
mvn install fails locally and in travis with this, but it gets the build
closer.  I'll need the closure team's help to figure out the rest.

After the changes discussed below, we get down to the Enforcer plugin
isn't able to find files in target/examples directory and there are
suspicious looking Soy errors in the test logs.

I'm at a loss.

What's been done so far:

  * bump the .travis.yml JVM to JDK8
    closure-template's pom.xml requires JDK8 (see the "1.8" in it).
    Unfortunately, OpenJDK 8 is not supported on TravisCI yet, so we
    have to switch to the Oracle JDK.

  * fix #46, the maven-compiler-plugin bug, by bumping to a working
 version of maven-compiler-plugin.