Automated [] rollback of changelist 130057364.
*** Reason for rollback ***

broke projects not using soy_js because they are sharing custom genrules with soy_js compilers

*** Original change description ***

migrate soy_js to use the open source compiler frontend.

The one wrinkle is that some projects explicitly pass the old default compiler,
I only found one example of this and am fixing it in [] but for now
there is explicit detection of this case and a workaround.

Also delete the gensoyjs.py file, i don't really know why this exists but on the cl that introduced it 5 years ago, the author said that this 'should be temporary'.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=130062699