soy: fix `mvn test` after a `mvn install`
This updates the POM to use the latest version of maven-compiler-plugin.

This addresses https://github.com/google/closure-templates/issues/46
running mvn compile twice without a clean causing duplicate classes
and fixes the travis-CI failure mode:
https://travis-ci.org/google/closure-templates/builds/146459924

"""
This seems to be straight up bug in maven 3.2 and 3.3. Related
tickets: https://issues.apache.org/jira/browse/MCOMPILER-235 and
https://issues.apache.org/jira/browse/MCOMPILER-236

A workaround is to add "false" to the maven-compiler-plugin. (This
seems to currently turn on Maven's maybe-not-so bright incremental
compiler. See the if-statement here and specifically the "stale
sources" code.
"""

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=128699856