Soy: Implement basic soy function parseFloat()
parseFloat() takes a single arg: the string to be parsed as a float.
If the string is not parseable, then the function returns null.

Note that partial numbers in JavaScript parse in parseFloat():
parseFloat('123abc') = 123
parseFloat('1.2.3') = 1.2

TODO: add documentation to
https://developers.google.com/closure/templates/docs/functions

OPEN_SOURCE_BREAKING_CHANGES=parseFloat() is a new core builtin. It may conflict with
user defined functions; the fix is to either rename the user-defined plugin or
migrate to this new builtin function.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=146392709