Fix 2 missingRequire bugs
1. we were failing to propagate requires metadata for the soy.$$registerDelegateFn call for deltemplates.
2. we were failing to propagate requires metadata the ordainer functions use by LetContentNode

This required some hackary to make Assignment work, this led me to the conclusion that Assignment probably shouldn't be a CodeChunk.WithValue, but rather if we want to consider an assignment/declaration to have a 'value' we should probably model it as an identifier with an initializing statement... which i guess would be a Composite?  something seems not quite right here.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=146680785