Rollforward []  which was rolled back in []  because it revealed a preexisting breakage in a project which was trying to enforce a particular namesapce structure.  The original cl caused additional files to be analyzed which violated the structure.  The errant project has been fixed.
*** Original change description ***

Fix some bugs in how we generate goog.requires for our jsdoc.

1. we failed to properly account for requires when composing aggregate types.  For example, 'html' would cause us to add a require for goog.soy.data.SanitizedContent, but list<html> would not.

2. we failed to add requires for all the relevant sanitized types.  The type signature for an 'html' field is '!goog.html.SafeHtml|!goog.soy.data.SanitizedHtml|!goog.soy.data.UnsanitizedText|string'
previously we would add a require for SanitizedHtml now we also add one for UnsanitizedText.  For the time being we do not do so for goog.html.SafeHtml, in order to prevent build breakages.

To support these more complex usecases i converted JsType to a final class with a builder, which required extracting a helper interface for the 'how to generate an assertion' logic.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=147538951