Remove SoyJsSrcOptions from the IncrementalDomSrc apis.  It is error prone.
Internally incrementaldomsrc uses SoyJsSrcOptions to configure parts of the jssrc backend (since they share code), however it needs SoyJsSrcOptions to be configured in a particular way.  This is error prone because the public apis accept SoyJsSrcOptions and don't validate it.

The solution is to change the public apis.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=147886927