Fix a bug where we failed to handle unknown types in proto init expressions prop[erly in the jbcsrc backend
The issue is that we would try to unbox a SoyExpression with unknown type as a
Message which isn't currently supported by the api.  The best long term
solution might be to provide a way to substitute a runtime type (e.g. pretend like
this is a foo.bar.Baz proto, instead of ?). But for now i simply took advantage
of the 'unboxUnchecked' api which doesn't have this problem (since it deals in
Expressions not SoyExpressions).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=150244508