Soy JS codegen: fix latent (unreported) goog.require bug.
Currently, SoyJsPluginUtils::applyDirective collects the symbols that need to be goog.require'd from the directive's arguments. But due to a spurious break statement, it stops collecting the symbols after the first arg that is not representable as a single expression.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=150330008