Soy: add+use CodeChunk.WithValue::initialStatements.
Currently, when a CodeChunk API has to interact with an API that only understands JsExprs, there is complicated logic involving CodeChunk::isRepresentableAsSingleExpr.

This CL makes isRepresentableAsSingleExpr package-private and introduces CodeChunk.WithValue::withInitialStatementsFrom() as a comprehensive API to "reassemble" a code chunk with specified dependencies.

The value of this CL is that it removes a few uses of the suspect statement() API, locks down isRepresentableAsSingleExpr, and should make future changes to the DSL easier.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=150472885