Soy: CodeChunk DSL: emancipate ConditionalBuilder from CodeChunk.Builder.
In the DSL, ConditionalBuilder is used to build up conditional expressions and statements. Currently, the only way to instantiate a ConditionalBuilder is through CodeChunk.Builder.

The connection between the two builder classes is not fundamental, and CodeChunk.Builder has confusing behavior. So I am going to get rid of it. This CL moves if_() to be a static method on CodeChunk and migrates existing callers.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=150688594