Enable the html parser for 'checking only'
This will cause us to run the pass unconditionally but not record the AST edits performed.  This will be a useful intermediate step to ensure that no new regressions are checked in while providing time to fix the tests that need tweaks to be compatible with the minor whitespace differences that will be introduced with the ast edits.

OPEN_SOURCE_BREAKING_CHANGES=strict soy templates with certain kinds of html
ambiguity will now be rejected by the compiler.  In particular
* all '<' characters need to be entity encoded if they aren't intended to begin a tag.  Previously we would allow such characters if they were immediately followed by whitespace
* all html tags, attributes and attribute values need to be opened and closed in the same control flow block
* contextually sensitive blocks (like <script>) need to have both the open and close tags in the same control flow block.
* html attributes need to be whitespace separated, even attributes following double quoted attribute values (e.g. a="a"b="b" will no longer be allowed)

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=150907256