Soy: CodeChunk DSL: change CodeChunk.WithValue::initialStatements to return an ImmutableSet.
Currently, the DSL uses CodeChunk.WithValue::isRepresentableAsSingleExpr as a heuristic for determining when it is safe to compose certain kinds of chunks. For example, if we are trying to take the logical and of two chunks A and B, and B has initial statements, the DSL will use an if-else statement (instead of `&&`) to ensure that B's initial statements are executed only if A is true.

But this heuristic is too conservative. In the above case, it is fine to use `&&` as long as B's initial statements are a subset of A's initial statements (which are always executed).

This CL changes initialStatements() to return an ImmutableSet. This will allow callers to test subset relationships of two chunks' initial statements so that the heuristics for &&, ||, and ?: expressions can be relaxed in the future. Incidentally, it also marks most of the DSL as @Immutable.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=151028046