Eagerly match the tags once we finish visiting IfNode or SwitchNode.
Without this change, the simple case will fail since the order has been messed up.
This change does change the error message a little bit, but I believe it is reasonable.

The approach we discussed offline (using a single stack) won't work for the cases where we want to match common suffix:

<div>
{if $foo}</div>{else}</div>{/if}

and also does not work for the following cases:

{if $foo}<div>{else}<div>{/if}
{if $foo}</div>{else}</div>{/if}

When we visit a close tag, we cannot report an error if the openTagStack is empty. But if both openTagStack and closeTagQueue are not empty but the top do not match, it is safe to report an error.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=151365878