Move the Desugaring pass to the end of the single file passes in the compiler
The main change is in the type heirarchy of the new nodes.  In several places in the compiler 'BlockNode' is used to mean 'introduces a new variable scope'.  However, the HtmlOpenTagNode et. al. was using it as a shorthand for ParentSoynode<StandaloneNode>.  This cl changes the type hierarchy and tweaks the definition of StandaloneNode so that StandaloneNodes don't have to be direct children of BlockNodes.

The second change is that the HtmlRewritePass now adds the 'tag name' as a child of the node (the first child).  This ensures that the tag name is visible to other passes that generically traverse the AST.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=151640270