Fix a bug in the autoescaper.
When attempting to match the end of an rcdata area, we need to find the close tag that matches the rcdata open tag.  however if there are other tag like things in the rcdata section we shouldn't give up.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=153886326