Rewrite the csp nonce injection pass in terms of the html nodes
This is necessary in order to break dependencies on the SlicedRawTextnodes
produced by the autoescaper.  It also yields some nice benefits in terms of
simplifying the implementation.

However, because the injection pass actually needs to modify the AST (and it depends on the new datastructures provided by the HtmlRewritePass) this can only be conditionally enabled.  However, once the autoescaper is rewritten to account for the html rewriting it means we won't break this pass

Also, as part of this i needed to move the CheckInvalidParamsVisitor to run _before_ the new 'injection' pass (so it wouldn't report violations on the compiler inserted code).  So that pass was simply folded into the new nonce injection pass.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=154060468