Soy: Remove CallNode() static factory methods
Make fields in CallNode itself mutable instead.

This is not ideal, but there is precedence for having mutable fields in nodes
(in CallNode, PrintNode, RawTextNode, and MsgFallbackGroupNode), and at least
some of it (the srcCalleeName stuff) is temporary until {call} parsing can be
migrated off of the nodes and into javacc.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=155029991