When extracting messages, sort them based on the source location instead of the id
This ordering is used to generate the xmb/xliff files, by using the source location instead of the id minor changes to messages will generate more stable diffs in the generated xmbs.  This will be useful for validating other parts of the compiler.

In addition to changing the ordering of messages in the extraction files, if there are multiple identical messages that differ only by description, the description in the extraction file may change due to the ordering change.  This behavior will not affect preexisting translations for the message (because neither message ids nor message contents depend on descriptions).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157049578