Rollforward [] which was rolled back in [] because it broke js compilation of projects with hyphenated tag names in messages.  This in turn revealed a slight incompatibility in the way we generate automatic placeholders which broke js compilation in a project.  The rollforward addresses this by enhancing the placeholder algorithm.
*** Original change description ***

Move parsing of MsgHtmlTagNodes into the HtmlRewritePass

The actual parsing logic is already there, so most of the work is:
* triggering it when in 'disabled' mode
* wrapping the tag nodes in the MsgPlaceholderNode and MsgHtmlTagNode
* changing MsgHtmlTagNode constructors
* keeping the same heuristics for deciding if two tags can use the same placeholder

This is not perfectly compatible due to the whitespace rules of the new parser.  You can see some examples of the incompatibilities in the MsgNodeTest

for example, given:
{msg desc="foo"}Hello<br />World<br/>{/msg}

Previous we would generate 2 placeholders, now we only produce 1 because we don't consider the whitespace in '<br />' to be meaningful.  To make sure that no one is relying on this i am diffing all xmbs in [].

Another incompatibility is that html tags inside of {msg} blocks are now
context sensitive, so
{let $t kind="text"}{msg desc="...}<a href="foo">hello</a>{/msg}{/let}

The tag-like things inside of the kind=text block are not parsed as
placeholders. To workaround users can wrap such messages in kind="html" blocks

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157493941