Migrate idom to use the new html parser and ast nodes
In order to do this i needed to add yet another PassManager option to enable the rewriting pass and disable the desugaring pass.  This is currently only set by the incremental dom compiler frontend.

The changes to the code generation backend are fairly trivial[0], most of the work is in the HtmlContextVisitor which enforces the various constraints of the idom backend.  Some of these can be relaxed in the future, but for now i continue to disallow dynamic attribute names, dynamic tag names and other similar things.

[0]: the new AST is more flexible and encodes things like attribute and tag names as children rather than node properties.  But other than those issues the basic structure of the code generation is maintained.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157497602