Automated [] rollback of changelist 157497602.
*** Reason for rollback ***

Breaks svg self closing nodes, CL with minimal reproducing test case:

[]

Sponge: []

This breaks the rendering of critical parts of material calendar web.

*** Original change description ***

Migrate idom to use the new html parser and ast nodes

In order to do this i needed to add yet another PassManager option to enable the rewriting pass and disable the desugaring pass.  This is currently only set by the incremental dom compiler frontend.

The changes to the code generation backend are fairly trivial[0], most of the work is in the HtmlContextVisitor which enforces the various constraints of the idom backend.  Some of these can be relaxed in the future, but for now i continue to di...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157578843