Rollforward of 157517369, the original CL had no specific concerns but it was preventing a clean rollback of a previous, and most likely orthogonal, CL.
Automated [] rollback of changelist 157576909.

*** Reason for rollback ***

Rollforward, this rollback was only required to make a clean rollback of [] There are no concerns with the code, but there was a risk that the two CLs were not orthogonal.

*** Original change description ***

Automated [] rollback of changelist 157517369.

*** Reason for rollback ***

To get a clean rollback of 157497602.

*** Original change description ***

Soy parser: Remove ExpressionParser and IN_CMD_TAG_UNSTRUCTURED

http://i.imgur.com/SAn3Nos.gif

***

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157617733