Move invocation of the autoescaper and optimizer into the PassManager
The reason this wasn't done before was because the autoescaper invalidates the invariant of most passes (it may create new templates).  This only happens for 'deprecated-contextual' templates, so the theory was always that as soon as deprecated-contextual was gone we could easily move it in.  It has been more than a year and things are unlikely to change in the next year.  Also, there is a need to run more compiler passes _after_ the autoescaper (notably the desugaring pass).  So to enable these new usecases the autoescaper invocation is being moved.  In particular:

* move invocation of the autoescaper and optimization passes from SoyFileSet -> PassManager
* create a new classification of CompilerFileSetPasses that run after the autoescaper
* create a configuration bit to disable the autoescaper (needed by incrementaldom and tests)
* along the way remove a bunch of guice injection points

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=157982725