Remove the genParamAlias method that was used to prevent soy from generating a local variable that conflicted with a js namespace or builtin function.
The original use cases have disapeared now that we have the
VarKey/UniqueNameGenerator system and the other usecases that have come up
appear to all be wrong. (we use it to mangle the name of the param in the
generated jsdoc types).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158213918