Automated [] rollback of changelist 158329624.
*** Reason for rollback ***

breaking people with the following code pattern
<script src="foo.js"/></script>

it appears that the autoescaper is getting confused by the self closing script tag, but the HtmlRewritePass doesn't so we are exposing some preexisting errors.

*** Original change description ***

Add two features to the htmlrewrite pass that are being backported from the autoescaper

1. add support for a legacy html feature for the rcdata tag <xmp>
2. report an error if we see an context changing close tag in PCDATA instead of the correct RCDATA state.

This should be compatible with the current autoescaper, though it may generate slightly different error messages (the source locations are definitely improved).

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158350199