Fix a test broken by []
The issue is tha we have a tracer that calls toString on a future.  The
implementation of AbstractFuture.toString() recenly changed to also call .get()
on the future.  Which breaks this test which tries to track exactly when .get()
is called.

The solution is to override toString() so we aren't sensitive to this.  I also
considered deleting the test.  I'm not sure how important it is given that tofu
is at EOL.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158807647