Make it possible to safely pass values from Soy into the Angular scope.
The safe pattern to pass a value is the following:
<div ng-init="a={$a}">The value of 'a' in the Angular scope: [[a]]</div>

The implementation makes ng-init a script attribute, and adds the characters
used in the usual Angular interpolation symbols (these are: {}[]) to the list
of symbols to escape in JS strings. This way, it's not possible to break out of
the string inside the Angular expression in ng-init, and it's not possible to
write an interpolation symbol without it being replaced by a hexadecimal JS
string escape.

Note that it's still possible to do it in an unsafe way - it just enables a
safe way to do it. It makes it possible to refactor mixed Soy - Angular
templates to be safe against direct Angular expression injections.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=158944709