Fix a bug in the DesugarHtmlNodesPass where we could change the semantics of an html tag in desugaring.
Given the following tag
  <div class=foo />

we would parse it as a self closing div tag with an unquoted attribute, however when we went to desugar it we would generate a raw text node that looks like
  <div class=foo/>

(note there is no space between foo and />)

This is actually a different tag!  this is an open div tag with an unquoted attribute value where the value is "foo/".

To account for this we detect whether or not a self closing tag possibly had a trailing unquoted attibute value and if it does, add a whitespace character.

A special case of this situation is <div/> a self closing tag with no attributes, this is not ambiguous and so we don't preserve the space character,  the only time it needs the whitespace is if the attribute ends in an unquoted attribute.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=159127115