Add basic support for Javascript Template Literals to the soy autoescaper.
Since template literals and js expressions can nest arbitrarily this required adding a counter to the Context object to track how deep we are.

We currently do not allow any soy commands to be interpolated in a js template literal.  Supporting this would not be very difficult, but it is left as future work since this is a very new and seldom used feature.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=159471688