For optional soy parameters make sure that the js record type explicitly adds a '|undefined' to the js type.
For params with explicit types like '{@param? optionalInt : int}' we would
already do this by marking the soy type as 'int|null' which in the jstype
system allows for undefined.  However, if the strict type was '?' we wouldn't
do this because in the soy type system '?|null' == '?'.  In jscompiler
parlance this isn't the same (if you are using --new_type_inf).  So instead for
optional untyped params explicitly add an '|undefined' to the generated type
signature.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=159767634