Have the jbcsrc compiler generate source locations for expressions
We didn't used to do this because the source locations for ExprNodes were innaccurate, but now that we have a unified parser this has been resolved.

To do this I added a .withSourceLocation method to Expression and SoyExpression.  This caused some difficulties with some of the subtypes of SoyExpression that specialize methods.  So i removed them.  The main loss from this was some code that theoretically optimized for cases were we call SoyExpression.box() followed by unbox, but beyond unit tests this never actually happens in the compiler.  So i removed that 'optimization'.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=160435015