Rollforward [] which was rolled back in [] because it caused stack overflow errors in templates which contained long javascript string literals with lots of escape sequences.  From what i can tell the issue was already present, but the previous cl made it slightly easier to trigger by inducing more backtracking in the regular expression evaluation. The fix is to make some of the repetitions 'possesive' which disables backtracking and thus disables recursion.
*** Original change description ***
Delete freshly dead code from the autoescaper.

Notably, the ContentSecurityPolicyPass (replaced by the ContentSecurityPolicyNonceInjectionPass), SlicedRawTextNodes (all uses have been migrated to the new AST nodes).  This also removes support for parsing html from the RawTextContextUpdater, which notably means modifying a lot of the regular expressions to remove various negative lookaheads.  Hopefully this is an overall complexity reduction and performance improvement in the RawTextContextUpdater.

Most of the work here is migrating some of the assertions in RawTextContextUpdater test to use a new testing framework since a lot of the things being validated by the test are now parts of other parts of the compiler which invalidated the old testing strategy.  Tests using this new framework are written in InferenceEngineTest

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=160601232