Fix some issues around handling null values in jbcsrc.
In jbcsrc we represent soy null as java null, however we also allow for interop with tofu soy plugins which like to consume and hand out NullData for null.  To handle this we map null -> NullData when calling functions however we were failing to map NullData->null when plugins returned values.

Fixing this exposed another more interesting issue involving how we optimize certain let expressions.

Normally when soy compiles a let expression we generate a subclass of SoyValueProvider so that we can evaluate it lazily.  however, in certain cases this is really really wasteful so we optimize it.  For example,

{let $foo : 2 /}

In this case we compile to IntegerData.forValue(2) which already is-a SoyValueProvider, so we are good to go.  However, we had a bug when the value in question was nullable.  For example,
{let $foo : randomInt(2) < 1 ? 1 : null /}

This is still considered optimizable, so instead of doing lazy evaluation we would evaluate it eagerly to something like
randomInt(2) < 1 ? IntegerData.forValue(1) : null

since that is what the boxing operation would do.  But we didn't want null for our SoyValueProvider instance.  The solution is to box null - > Runtime.NULL_PROVIDER which was intended for this usecase.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=161000580