jbcsrc: add a test to make sure we are annotating our gencode with line numbers correctly
The strategy is to instrument the integration tests with calls to a special plugin function with a testonly compiler pass.  the pass will rewrite

$foo

to

inspectSourceLocation($foo, 'bar.soy', 2)

where bar.soy is the filename and 2 is the line number.  Then at runtime the
function will inspect its own stack to make sure that the caller is at that
line number according to the jvm.  In theory we should also instrument SoyNodes
(like if-statements) to make sure they get correct source locations, but for
now we are skipping it since expressions tend to be more interesting.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=161005261