Refactor Field/FieldInterpreter so that they can be constructed without a type registry.
The logic for calculating types and converting values is still unified to avoid
duplicating the logic for traversing and interpreting proto descriptors. (in the past this was split which lead to bugs and inconsistencies).

This will pave the way for eliminating SoyCustomValueConverter and
SoyTypeProvider as well as eliminating the various complex guice apis.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=161226511