Add a code chunk api for function literals and use it for template declarations.
This changes the way we declare functions for incremental dom from

    function foo(...){}

to

    var foo = function...){}

which should be identical in behavior.

This also ran into a bunch of issues with visitChildrenReturningCodeChunk where extra newlines, indents and semi colons were added.  Some of this was a preexisting issue (as demonstrated by some of the golden file tests). I have added additional workarounds for this in LeafStatement, but the situation is still a mess.

Is there an overarching strategy to who is responsible for terminating statements with semi colons/newlines?  everything seems pretty ad-hoc.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=161446074