Fix a quadratic performance problem in the CombineConsecutiveRawTextNodesVisitor
This is a classic N^2 algorithm where we were essentially doing
String out = ""
for (String s : strings) {
  out +=  s;
}

The fix is to measure and then concatenate the strings.  This is complicated slightly by the fact that these are RawTextNodes which in addition to text also contain a set 'offsets' which record where there are 'source discontinuities' in the text.  So we need to concatenate that data as well.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=162288912