Add support for warnings to soy
Warnings should be used sparingly but they will be useful for deprecation/migration efforts to 'stop the bleeding'.

The most controversial part of this is probably the 'isWarning' method on SoyError.  I played around with introducing a parallel type for this 'SoyWarning' but decided against it.  SoyError is already a public api and as such things like SoyCompilationException (which is also public) needs to accept them.  Adding a parallel type would mean parallel methods/fields/parameters in a number of places which seems unnecessarily complex.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=162304034