Add compiler warnings for the {css and {xid commands, users should migrate to the function syntax instead.
Due to pervasive use of the exploding error reporter in tests, this caused a number of unit tests to fail.  I migrated some of them to use the css/xid functions and in other cases I simply deleted the tests.  This does mean there will be a reduction in test coverage for these deprecated statements but i think this is reasonable because we still have integrationtest coverage and these things are on the path to destruction.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=162536501