Change the way we pop optional tags in StrictHtmlValidationPass. This should improve the error message for mismatch optional tags, and also introduce more strict validation for the content model (for example, report an error for <div><li></div> since <li> is not allowed to be closed by </div>).
Before this CL, since we don't support full content model, we used a heuristic for optional tags: whenever we try to match a close tag, we unconditionally pop out all optional tags that are not matching. This sometimes corrupts the matching algorithm and produces bad error messages that are far from the actual mismatch tags.

In this CL, we introduce a rule that an optional open tag can only be popped by certain close tags. This should prevent us to pop more than we expected and corrupt the tag stack.

This implements half of the content model described in https://www.w3.org/TR/html5/syntax.html#optional-tags. The remaining half is whether we should pop optional tags when we see open tags; we will revisit that in the future since it will require major changes to the way we run the compiler pass.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=162836721