Automated [] rollback of changelist 163737652.
*** Reason for rollback ***

Breaks projects explicitly passing [] to their af_soy_options rule

they are probably doing this because it should have been part of the common set

*** Original change description ***

Fix a bug in proto init for jbcsrc.

The bug was that when assigning a soy value to a safe proto field (like SafeHtmlProto) we would need unpack the SanitizedContent object into a proto first.  This is accomplished by invoking the corresponding methods on SanitizedContent (like toSafeHtmlProto).  However, if the value being assigned was untyped for whatever reason we would fail to insert a cast operator which would cause us to generate bad code (essentially trying to call .toSafeHtmlProto on a g...

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=163748326