Rollforward [] which was rolled back in [] because it created duplicate dep errors in some af_soy_options rules.  These duplicates were independently addressed in []
*** Original change description ***
Fix a bug in proto init for jbcsrc.

The bug was that when assigning a soy value to a safe proto field (like SafeHtmlProto) we would need unpack the SanitizedContent object into a proto first.  This is accomplished by invoking the corresponding methods on SanitizedContent (like toSafeHtmlProto).  However, if the value being assigned was untyped for whatever reason we would fail to insert a cast operator which would cause us to generate bad code (essentially trying to call .toSafeHtmlProto on a generic SoyValue object).  This in turn leads to verification errors in the bytecode.

The solution is to insert a cast.

While doing this i noticed a few more weird things with how the proto init code was generating cast operators and fixed them as well.

***

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=163781446