Add the initial logging apis for runtime rendering.
SoyLogger: The interface that will be implemented by users wishing to process
log statements.
LoggingAdvisingAppendable: An internal object to collecting both append and
logging operations.  This has also defined a number of implementations but for
all practical purposes it is identical to AdvisingAppendable in how it is
currently being used.
LogStatement: A value type that will be generated whenever a foolog statement
is encountered
LoggingFunctionInvocation: A value type that will be generated whenever a
logging function is invoked.

This also modifies the SoyValue/SoyValueProvider apis to work in terms of these
new types instead of AdvisingAppendable.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=164490868