Add support for logonly=true in the soysauce backend
Since the decision to make something logonly is data dependent the easiest approach was just to have the output enforce the policy.  The main complexity is that we need to efficiently track when to disable logonly which is non-trivial due to the recursive nature of   To accomplish this i used a bitset to encode a stack of  on/off booleans and a counter to see how many logonly elements we are nested within.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=165024877